requirements.txt
python-docx
langchain-ollama
pillow
deep-translator
langdetect